requirements.txt
tensorflow>=2.2.0
numpy~=1.18.5
pandas~=1.0.5
Pillow~=7.0.0
matplotlib~=3.2.2
deepface~=0.0.33
autocrop~=1.0.2
scikit-learn~=0.23.1